requirements.txt
allennlp

# Pin these to avoid conflicts.
boto==2.49.0
boto3==1.9.115
botocore==1.12.115

# Registering pages.
registrable

# Webserver.
gevent==1.3.6
flask>=1.0.2,<2.0.0
flask-cors>=3.0.7,<4.0.0
flask-login>=0.4.1,<1.0.0
flask-caching==1.7.2
gunicorn>=19.9.0,<20.0.0
dash==1.5.1,<2.0.0
dash-daq==0.1.4,<1.0.0
dash-bootstrap-components==0.7.2

# Data classes.
attrs>=18.2.0,<19.0.0
cattrs==0.9.0

# Serializing to/from toml files.
toml>=0.10.0,<1.0.0

# Command line interface.
click>=6.7,<7.0
click-completion>=0.5.0,<1.0.0
colorama>=0.3.9,<1.0.0

# Querying GPU info.
nvidia-ml-py3==7.352.0
py3nvml==0.2.5
